+++ project_backend/Interfaces/IUserRepository.cs
@@ -2,10 +2,10 @@
 {
     public interface IUserRepository
     {
-        public int AddUserBalance(int userId, int balance);
-        public int GetUserBalance(int userId);
-        public void UpdateUserBalance(int userId, int reducedBalance);
-        public void AppendPurchaseHistory(int userId, string itemName, int quantity, int unitPrice);
+        public decimal AddUserBalance(int userId, decimal balance);
+        public decimal GetUserBalance(int userId);
+        public void UpdateUserBalance(int userId, decimal reducedBalance);
+        public void AppendPurchaseHistory(int userId, string itemName, int quantity, decimal unitPrice);
 
     }
 }
